Moved settings button to it's own component
